backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.3.1
ollama==0.1.7
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
werkzeug==3.0.1
